pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude_data_agent"
version = "0.1.0"
description = "Data agent for processing and analyzing data using Claude"
requires-python = ">=3.10"
dependencies = [
    # Core data science libraries
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    # Time series analysis
    "statsmodels>=0.14.0",
    "plotly>=5.15.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    # Essential causal inference (validated working versions)
    "econml>=0.13.0",
    "linearmodels>=4.25.0",
    "arch>=5.3.0",
    # Network analysis and optimization
    "networkx>=2.8.0",
    "cvxpy>=1.3.0",
    # Time series advanced methods
    "ruptures>=1.1.0",
    "prophet>=1.1.0",
    # Machine learning for clustering
    "umap-learn>=0.5.0",
    "hdbscan>=0.8.0",
    # Anomaly detection and outlier analysis
    "pyod>=1.0.0",
    # Extreme value theory
    "pyextremes>=2.2.0",
    # Optimal transport
    "POT>=0.8.0",
    # Geospatial analysis for pipeline networks
    "geopandas>=0.13.0",
    # High performance computing
    "numba>=0.56.0",
    "dask>=2022.0.0",
    # LLM libraries
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "tiktoken>=0.5.0",
    "litellm>=1.0.0",
    # Data validation and structures
    "pydantic>=2.0.0",
    # Environment and configuration
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "jinja2>=3.0.0",
    # Progress tracking and logging
    "tqdm>=4.65.0",
    "rich>=13.0.0",
    # Version control integration
    "gitpython>=3.1.0",
    # File handling
    "openpyxl>=3.1.0",
    "xlsxwriter>=3.1.0",
    # Vector store and document processing
    "qdrant-client>=1.6.0",
    "pypdf>=4.0.0",
    "beautifulsoup4>=4.12.0",
    "selenium>=4.15.0",
    "markdownify>=0.11.0",
    "requests>=2.31.0",
    "claude-code-sdk>=0.0.23",
    "structlog>=23.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]

[tool.setuptools]
py-modules = ["run_agent", "runner_utils"]

[tool.ruff]
target-version = "py310"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["F401", "F811"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "plotly.*",
    "statsmodels.*",
]
ignore_missing_imports = true
